# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/words:
    get:
      operationId: getAdminWords
      tags: [admin]
      summary: List words learned by the model
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: Learned vocabulary with teaching attribution
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminLearnedWord"
  /admin/words/{word}:
    delete:
      operationId: deleteAdminWord
      tags: [admin]
      summary: Delete a learned word and its model connections
      parameters:
        - name: word
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted learned word
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminWordDeletion"
        "400":
          description: Invalid learned word
        "404":
          description: Learned word not found
...
  /admin/chats/{chatId}:
    get:
      operationId: getAdminChat
      tags: [admin]
      summary: Get an AI-involving chat for administrators
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI conversation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminChatDetail"
        "404":
          description: AI conversation not found
  /admin/accounts/{username}/ban:
    post:
      operationId: banAdminAccount
      tags: [admin]
      summary: Ban an account and delete its saved chat history
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Banned account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
        "403":
          description: Self-ban or administrator-ban rejected
        "404":
          description: Account not found
  /admin/admins:
    post:
      operationId: grantAdmin
      tags: [admin]
      summary: Grant administrator access to an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminGrantAdminInput"
      responses:
        "200":
          description: Account with administrator access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
        "400":
          description: Invalid username
        "404":
          description: Account not found
components:
  schemas:
...
    ChatParticipant:
      type: object
      required: [username, displayName, isBrain]
      properties:
        username:
          type: string
        displayName:
          type: string
        isBrain:
          type: boolean
...
    AdminAccount:
      type: object
      required: [username, createdAt, isAdmin, isBanned]
      properties:
        username:
          type: string
        createdAt:
          type: string
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
    AdminChatMessage:
      type: object
      required: [id, sender, content, createdAt]
      properties:
        id:
          type: string
        sender:
          $ref: "#/components/schemas/ChatParticipant"
        content:
          type: string
        createdAt:
          type: string
    AdminChatSummary:
      type: object
      required: [id, type, title, participants, includeBrain, createdAt, updatedAt, lastMessage]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [direct, group]
        title:
          type: string
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ChatParticipant"
        includeBrain:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        lastMessage:
          oneOf:
            - $ref: "#/components/schemas/AdminChatMessage"
            - type: "null"
    AdminChatDetail:
      allOf:
        - $ref: "#/components/schemas/AdminChatSummary"
        - type: object
          required: [messages]
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/AdminChatMessage"
    AdminGrantAdminInput:
      type: object
      required: [username]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: "^[A-Za-z0-9_-]+$"
    AdminLearnedWordTeacher:
      type: object
      required: [username, count]
      properties:
        username:
          type: string
        count:
          type: number
          minimum: 0
    AdminLearnedWord:
      type: object
      required: [word, count, taughtBy]
      properties:
        word:
          type: string
        count:
          type: number
          minimum: 0
        taughtBy:
          type: array
          items:
            $ref: "#/components/schemas/AdminLearnedWordTeacher"
    AdminWordDeletion:
      type: object
      required: [word, deleted, remainingWords]
      properties:
        word:
          type: string
        deleted:
          type: boolean
        remainingWords:
          type: number
          minimum: 0